Move touchRating(position:configuration:) from SBTouchHandler.swift to SBStarRatingControl.swift and rename as touchRating(position:)
